+++ src/main/java/com/permitseoul/permitserver/domain/reservation/core/repository/ReservationRepository.java
@@ -4,9 +4,10 @@ import com.permitseoul.permitserver.domain.reservation.core.domain.entity.Reserv
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
 public interface ReservationRepository extends JpaRepository<ReservationEntity, Long> {
-    Optional<ReservationEntity> findByOrderIdAndTotalAmountAndUserId(final String orderId, int totalAmount, long userId);
+    Optional<ReservationEntity> findByOrderIdAndTotalAmountAndUserId(final String orderId, BigDecimal totalAmount, long userId);
 }
